move logic to uistate store
