+++ client/src/pages/src/visualizer-component/parser/parser.ts
@@ -1,14 +1,9 @@
 import { UserAnnotation } from '../types/annotationType';
 import { BackendState } from '../types/backendType';
 import { GenericGraph } from '../types/frontendType';
-import { UiState } from '../types/uiState';
 
 export interface Parser {
-  parseInitialState: (
-    backendStructure: BackendState,
-    annotation: UserAnnotation,
-    uiState: UiState
-  ) => GenericGraph;
+  parseInitialState: (backendStructure: BackendState, annotation: UserAnnotation) => GenericGraph;
 }
 
 /**
