+++ client/src/pages/src/visualizer-component/types/uiState.ts
@@ -1,18 +1,17 @@
-import { create } from 'zustand';
 import { UserAnnotation } from './annotationType';
-
-export enum VisualizerType {
-  LINKED_LIST = 'LINKED_LIST',
-  BINARY_TREE = 'BINARY_TREE',
-  GRAPH = 'GRAPH',
-  ARRAY = 'ARRAY',
-}
+import { VisualizerComponent } from '../visulizer/visualizer';
+import { Parser } from '../parser/parser';
+import { VisualizerType } from './visaulizerType';
+import { visualizerFactory } from '../visulizer/visualizerFactory';
+import { parserFactory } from '../parser/parserFactory';
 
 export type UiState = {
   visualizerType: VisualizerType;
   width: number;
   height: number;
   userAnnotation: UserAnnotation;
+  visComponent: VisualizerComponent;
+  parser: Parser;
 };
 
 export const DEFAULT_UISTATE: UiState = {
@@ -23,6 +22,8 @@ export const DEFAULT_UISTATE: UiState = {
     stackAnnotation: {},
     typeAnnotation: {},
   },
+  visComponent: visualizerFactory(VisualizerType.LINKED_LIST),
+  parser: parserFactory(VisualizerType.LINKED_LIST),
 };
 
 export const NODE_SIZE = 30;
