+++ client/src/pages/src/visualizer-component/types/visaulizerType.ts
@@ -1,0 +1,6 @@
+export enum VisualizerType {
+  LINKED_LIST = 'LINKED_LIST',
+  BINARY_TREE = 'BINARY_TREE',
+  GRAPH = 'GRAPH',
+  ARRAY = 'ARRAY',
+}
